+++ lib/cuprum/operation.rb
@@ -93,11 +93,6 @@ module Cuprum
         called? ? result.failure? : false
       end # method success?
 
-      # @return [Boolean] true if the most recent was halted, otherwise false.
-      def halted?
-        called? ? result.halted? : false
-      end # method halted?
-
       # Clears the reference to the most recent call of the operation, if any.
       # This allows the result and any referenced data to be garbage collected.
       # Use this method to clear any instance variables or state internal to the
@@ -136,9 +131,6 @@ module Cuprum
 
     # @!method failure?
     #   (see Cuprum::Operation::Mixin#failure?)
-
-    # @!method halted?
-    #   (see Cuprum::Operation::Mixin#halted?)
 
     # @!method reset!
     #   (see Cuprum::Operation::Mixin#reset!)